Warp Python/requirements.txt
warp-lang
pillow
pywin32
